requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
captcha==0.3
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.3
ffmpeg==1.4
idna==2.9
multidict==4.7.5
Pillow==7.1.2
pycares==3.1.1
pycparser==2.20
websockets==8.1
yarl==1.4.2
youtube-dl==2020.5.8
